requirements.txt
flask==1.1.0
flask_restful==0.3.8
ruamel.yaml==0.16.0
#numpy==1.13.1
#matplotlib==2.0.2
unittest-xml-reporting==2.5.1
coverage==4.5.4
nose==1.3.7
nose-htmloutput==0.6.0